Update on "[NNC Testing] Randomized loop nest infrastructure"
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D33234529](https://our.internmc.facebook.com/intern/diff/D33234529)

[ghstack-poisoned]
