+++ torch/csrc/jit/tensorexpr/loopnest.h
@@ -558,7 +558,14 @@ class TORCH_API LoopNest {
   }
   std::vector<BufPtr> getIntermediateBufs() const;
 
-  void randomTransform(int64_t seed);
+  // Finds which is the outer for loop between a and b for loops. Returns
+  // nullptr if they don't have same parent.
+  static ForPtr findOuterFor(ForPtr a, ForPtr b);
+
+  // Applies a series of loop optimizations chosen randomly. This is only for
+  // testing purposes. This allows automatic stress testing of NNC loop
+  // transformations.
+  void loopnestRandomization(int64_t seed);
 
  private:
   void initialize(
